requirements/development.txt
-r base.txt
fastapi[standard]
strawberry-graphql[fastapi]